requirements.txt
streamlit
requests
python-dateutil
pydantic
google-genai
